refactor common wrappers for tar and zip to transmat.util
