+++ cmd/rio/demuxes.go
@@ -5,8 +5,8 @@ import (
 
 	"go.polydawn.net/go-timeless-api/rio"
 	"go.polydawn.net/rio/transmat/git"
-	"go.polydawn.net/rio/transmat/tar"
-	"go.polydawn.net/rio/transmat/zip"
+	tartrans "go.polydawn.net/rio/transmat/tar"
+	ziptrans "go.polydawn.net/rio/transmat/zip"
 )
 
 func demuxPackTool(packType string) (rio.PackFunc, error) {
@@ -26,6 +26,8 @@ func demuxUnpackTool(packType string) (rio.UnpackFunc, error) {
 		return tartrans.Unpack, nil
 	case "git":
 		return git.Unpack, nil
+	case "zip":
+		return ziptrans.Unpack, nil
 	default:
 		return nil, Errorf(rio.ErrUsage, "unsupported packtype %q", packType)
 	}
@@ -35,6 +37,8 @@ func demuxScanTool(packType string) (rio.ScanFunc, error) {
 	switch packType {
 	case "tar":
 		return tartrans.Scan, nil
+	case "zip":
+		return ziptrans.Scan, nil
 	default:
 		return nil, Errorf(rio.ErrUsage, "unsupported packtype %q", packType)
 	}
@@ -44,6 +48,8 @@ func demuxMirrorTool(packType string) (rio.MirrorFunc, error) {
 	switch packType {
 	case "tar":
 		return tartrans.Mirror, nil
+	case "zip":
+		return ziptrans.Mirror, nil
 	default:
 		return nil, Errorf(rio.ErrUsage, "unsupported packtype %q", packType)
 	}
