+++ transmat/tar/doc.go
@@ -5,7 +5,15 @@
 package tartrans
 
 import (
-	"go.polydawn.net/go-timeless-api"
+	api "go.polydawn.net/go-timeless-api"
+	"go.polydawn.net/go-timeless-api/rio"
+	"go.polydawn.net/rio/transmat/util"
 )
 
 const PackType = api.PackType("tar")
+
+var (
+	Mirror rio.MirrorFunc = util.CreateMirror(unpackTar)
+	Scan   rio.ScanFunc   = util.CreateScanner(PackType, unpackTar)
+	Unpack rio.UnpackFunc = util.CreateUnpack(PackType, unpackTar)
+)
